Add doc versioning
The old URLs will automatically redirect to the latest version.
Versions are selectable in the left-side menu, and from a dropdown at the top of pages for versioned components.
If the user switches version from the dropdown, they'll be taken to the same page in the new version, or the index page if the same page doesn't exist.